+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All Notable changes to `flipboxdigital\craft-sortable-associations` will be documented in this file
 
+##Unreleased
+### Fixed
+- Saving an association record may trigger the save operation twice.
+
 ## 1.0.0 - 2018-04-23
 ### Changed
 - `SortableAssociations` service no longer uses a constant to identify the record alias.